Extracted common config to be used in both plugins
